requirements.txt
# Core libraries
opencv-python>=4.7.0.72
face-recognition>=1.3.0
dlib>=19.24.0
numpy>=1.24.0
Pillow>=9.4.0

# Audio processing
pyttsx3>=2.90
gTTS>=2.3.1
playsound>=1.3.0

# Data handling
pandas>=1.5.3
openpyxl>=3.1.1
google-api-python-client>=2.79.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Web dashboard
Flask>=2.2.3
Flask-Login>=0.6.2
Werkzeug>=2.2.3
Jinja2>=3.1.2
streamlit>=1.21.0  # Alternative to Flask

# Configuration and utilities
PyYAML>=6.0
python-dotenv>=1.0.0
colorlog>=6.7.0

# Testing and development
pytest>=7.3.1
pytest-mock>=3.10.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.2.0

# Documentation
Sphinx>=6.1.3
sphinx-rtd-theme>=1.2.0

